requirements.txt
aws-cdk-lib==2.126.0
constructs>=10.0.0,<11.0.0
PyYAML==6.0.1
requests==2.31.0
aws-cdk.lambda-layer-kubectl-v28